% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{.axis_active}
\alias{.axis_active}
\title{Decide whether a given axis should render}
\usage{
.axis_active(track, resolved_theme, side)
}
\arguments{
\item{track}{The \code{SeqTrackR6}.}

\item{resolved_theme}{Track resolved theme.}

\item{side}{One of \code{"x1"}, \code{"x2"}, \code{"y1"}, \code{"y2"}.}
}
\value{
Logical.
}
\description{
Decide whether a given axis should render
}
\keyword{internal}
