% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_ops.R
\name{.data_to_npc}
\alias{.data_to_npc}
\title{Normalise a data x value to the unit interval within a scale range}
\usage{
.data_to_npc(x, xscale)
}
\arguments{
\item{x}{Numeric vector of data x values.}

\item{xscale}{Numeric vector of length 2: \code{c(min, max)}.}
}
\value{
Numeric vector of the same length as \code{x}, clamped to \verb{[0, 1]}.
}
\description{
Normalise a data x value to the unit interval within a scale range
}
\keyword{internal}
