% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_ops.R
\name{.data_to_npc_y}
\alias{.data_to_npc_y}
\title{Normalise a data y value to the unit interval within a scale range}
\usage{
.data_to_npc_y(y, yscale)
}
\arguments{
\item{y}{Numeric vector of data y values.}

\item{yscale}{Numeric vector of length 2: \code{c(min, max)}.}
}
\value{
Numeric vector of the same length as \code{y}, clamped to \verb{[0, 1]}.
}
\description{
Normalise a data y value to the unit interval within a scale range
}
\keyword{internal}
