% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{.draw_track_axes}
\alias{.draw_track_axes}
\title{Draw all active axes for one track}
\usage{
.draw_track_axes(track, panels)
}
\arguments{
\item{track}{A \code{SeqTrackR6}.}

\item{panels}{Panel list (from \code{layoutGrid()}).}
}
\value{
Invisibly \code{NULL}.
}
\description{
Draw all active axes for one track
}
\keyword{internal}
