% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{.resolve_bin_fun}
\alias{.resolve_bin_fun}
\title{Resolve the \code{fun} argument to a scalar aggregation function}
\usage{
.resolve_bin_fun(fun)
}
\description{
Resolve the \code{fun} argument to a scalar aggregation function
}
\keyword{internal}
