% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_scalebar.R
\name{.scalebar_auto_label}
\alias{.scalebar_auto_label}
\title{Auto-format a scalebar label from a base-pair length}
\usage{
.scalebar_auto_label(bp)
}
\arguments{
\item{bp}{Numeric. Length in bp.}
}
\value{
Character label such as \code{"50 kb"}.
}
\description{
Auto-format a scalebar label from a base-pair length
}
\keyword{internal}
