% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_chip.R
\name{.split_gr_by_sample}
\alias{.split_gr_by_sample}
\title{Split a GRanges by sample column into a named list}
\usage{
.split_gr_by_sample(gr, sample_col)
}
\description{
Split a GRanges by sample column into a named list
}
\keyword{internal}
