% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{.stop_if_not_granges}
\alias{.stop_if_not_granges}
\title{Assert that an object is a GRanges}
\usage{
.stop_if_not_granges(x, arg_name)
}
\arguments{
\item{x}{Object to check.}

\item{arg_name}{Name of the argument, used in the error message.}
}
\value{
Invisibly \code{TRUE} if the check passes; otherwise stops.
}
\description{
Internal helper. Errors if \code{x} does not inherit from \code{"GRanges"}.
}
\keyword{internal}
