% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{flexoki_palette}
\alias{flexoki_palette}
\title{Load colors from Kepano's Flexoki palette}
\usage{
flexoki_palette(n, shade = 400)
}
\arguments{
\item{n}{Number of colors to be produced. If > 9, an equally distributed
color ramp will be produced (and probably look bad).}

\item{shade}{Integer shade level. One of \code{200}, \code{300}, \code{400}, or \code{600}.}
}
\value{
Vector of hex colors.
}
\description{
Loads colors from Kepano's Flexoki palette.
}
