% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{is_seq_file_conn}
\alias{is_seq_file_conn}
\title{Test whether \code{x} is a SeqPlotR file connection}
\usage{
is_seq_file_conn(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if \code{x} inherits from \code{SeqFileConn}, otherwise \code{FALSE}.
}
\description{
Connection objects from \code{\link[=open_bigwig]{open_bigwig()}}, \code{\link[=open_bam]{open_bam()}}, \code{\link[=open_hic]{open_hic()}},
and \code{\link[=open_h5]{open_h5()}} all carry the \code{SeqFileConn} class.
}
