% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{is_seq_legend_spec}
\alias{is_seq_legend_spec}
\title{Test if an object is a SeqLegendSpec}
\usage{
is_seq_legend_spec(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
Logical scalar.
}
\description{
Test if an object is a SeqLegendSpec
}
