% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{op-null-coalesce}
\alias{op-null-coalesce}
\alias{\%||\%}
\title{Null-coalescing operator}
\usage{
e1 \%||\% e2
}
\arguments{
\item{e1}{Left-hand side value.}

\item{e2}{Right-hand side fallback value.}
}
\value{
\code{e1} if not \code{NULL}, else \code{e2}.
}
\description{
Returns \code{e1} if it is not \code{NULL}, otherwise returns \code{e2}.
}
\examples{
NULL \%||\% "fallback"
"a" \%||\% "b"

}
