% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{seq_blank}
\alias{seq_blank}
\title{Return a blank plot-level placeholder}
\usage{
seq_blank()
}
\value{
A list with class \code{"SeqBlank"}.
}
\description{
Used to occupy a named cell in a patchwork layout without rendering anything.
}
