% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_area.R
\name{SeqAreaR6}
\alias{SeqAreaR6}
\title{SeqArea R6 class}
\description{
SeqArea R6 class

SeqArea R6 class
}
\details{
Internal R6 generator backing \code{\link[=seq_area]{seq_area()}}. Inherits from \code{\link{SeqElementR6}}.
}
\keyword{internal}
\section{Super class}{
\code{SeqPlotR::SeqElement} -> \code{SeqArea}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SeqArea-new}{\code{SeqAreaR6$new()}}
\item \href{#method-SeqArea-prep}{\code{SeqAreaR6$prep()}}
\item \href{#method-SeqArea-draw}{\code{SeqAreaR6$draw()}}
\item \href{#method-SeqArea-clone}{\code{SeqAreaR6$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SeqPlotR" data-topic="SeqElement" data-id=".infer_scale_y"><a href='../../SeqPlotR/html/SeqElement.html#method-SeqElement-.infer_scale_y'><code>SeqPlotR::SeqElement$.infer_scale_y()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqPlotR" data-topic="SeqElement" data-id="collect_legend_keys"><a href='../../SeqPlotR/html/SeqElement.html#method-SeqElement-collect_legend_keys'><code>SeqPlotR::SeqElement$collect_legend_keys()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqPlotR" data-topic="SeqElement" data-id="resolve"><a href='../../SeqPlotR/html/SeqElement.html#method-SeqElement-resolve'><code>SeqPlotR::SeqElement$resolve()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqArea-new"></a>}}
\if{latex}{\out{\hypertarget{method-SeqArea-new}{}}}
\subsection{Method \code{new()}}{
Construct a SeqAreaR6.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqAreaR6$new(data = NULL, mapping = NULL, aesthetics = aes(), ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Optional \code{GRanges}.}

\item{\code{mapping}}{Optional \code{SeqMap}. Required: \code{x}, \code{y}.}

\item{\code{aesthetics}}{Optional \code{SeqAes}. \code{baseline} sets the closing
y-value (default 0).}

\item{\code{...}}{Unused.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqArea-prep"></a>}}
\if{latex}{\out{\hypertarget{method-SeqArea-prep}{}}}
\subsection{Method \code{prep()}}{
Resolve mappings, clip to windows, transform to npc, then
build a closed polygon by appending a reversed baseline path.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqAreaR6$prep(layout_track, track_windows)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{layout_track}}{Per-window panel metadata list.}

\item{\code{track_windows}}{Track's \code{windows} GRanges.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqArea-draw"></a>}}
\if{latex}{\out{\hypertarget{method-SeqArea-draw}{}}}
\subsection{Method \code{draw()}}{
Draw filled polygons via \code{grid::grid.polygon()}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqAreaR6$draw()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqArea-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SeqArea-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqAreaR6$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
