% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_bar.R
\name{SeqBarR6}
\alias{SeqBarR6}
\title{SeqBar R6 class}
\description{
SeqBar R6 class

SeqBar R6 class
}
\details{
Internal R6 generator backing \code{\link[=seq_bar]{seq_bar()}}. Inherits from \code{\link{SeqElementR6}}.
}
\keyword{internal}
\section{Super class}{
\code{SeqPlotR::SeqElement} -> \code{SeqBar}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SeqBar-new}{\code{SeqBarR6$new()}}
\item \href{#method-SeqBar-prep}{\code{SeqBarR6$prep()}}
\item \href{#method-SeqBar-draw}{\code{SeqBarR6$draw()}}
\item \href{#method-SeqBar-clone}{\code{SeqBarR6$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SeqPlotR" data-topic="SeqElement" data-id=".infer_scale_y"><a href='../../SeqPlotR/html/SeqElement.html#method-SeqElement-.infer_scale_y'><code>SeqPlotR::SeqElement$.infer_scale_y()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqPlotR" data-topic="SeqElement" data-id="collect_legend_keys"><a href='../../SeqPlotR/html/SeqElement.html#method-SeqElement-collect_legend_keys'><code>SeqPlotR::SeqElement$collect_legend_keys()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqPlotR" data-topic="SeqElement" data-id="resolve"><a href='../../SeqPlotR/html/SeqElement.html#method-SeqElement-resolve'><code>SeqPlotR::SeqElement$resolve()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqBar-new"></a>}}
\if{latex}{\out{\hypertarget{method-SeqBar-new}{}}}
\subsection{Method \code{new()}}{
Construct a SeqBarR6.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqBarR6$new(data = NULL, mapping = NULL, aesthetics = aes(), ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Optional \code{GRanges}.}

\item{\code{mapping}}{Optional \code{SeqMap}. Recognised: \code{x}, \code{y}, \code{group}, \code{fill}.}

\item{\code{aesthetics}}{Optional \code{SeqAes}. Supports \code{fill}, \code{color},
\code{linewidth}, \code{alpha}.}

\item{\code{...}}{Unused.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqBar-prep"></a>}}
\if{latex}{\out{\hypertarget{method-SeqBar-prep}{}}}
\subsection{Method \code{prep()}}{
Resolve mappings, clip to windows, transform to npc. When
a \code{group} mapping is present, bars stack at identical x positions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqBarR6$prep(layout_track, track_windows)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{layout_track}}{Per-window panel metadata list.}

\item{\code{track_windows}}{Track's \code{windows} GRanges.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqBar-draw"></a>}}
\if{latex}{\out{\hypertarget{method-SeqBar-draw}{}}}
\subsection{Method \code{draw()}}{
Draw bars via \code{grid::grid.rect()}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqBarR6$draw()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqBar-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SeqBar-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqBarR6$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
