% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_curve.R
\name{SeqCurveR6}
\alias{SeqCurveR6}
\title{SeqCurve R6 class}
\description{
SeqCurve R6 class

SeqCurve R6 class
}
\details{
Internal R6 generator backing \code{\link[=seq_curve]{seq_curve()}}. Inherits from \code{\link{SeqElementR6}}.
}
\keyword{internal}
\section{Super class}{
\code{SeqPlotR::SeqElement} -> \code{SeqCurve}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SeqCurve-new}{\code{SeqCurveR6$new()}}
\item \href{#method-SeqCurve-prep}{\code{SeqCurveR6$prep()}}
\item \href{#method-SeqCurve-draw}{\code{SeqCurveR6$draw()}}
\item \href{#method-SeqCurve-clone}{\code{SeqCurveR6$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SeqPlotR" data-topic="SeqElement" data-id=".infer_scale_y"><a href='../../SeqPlotR/html/SeqElement.html#method-SeqElement-.infer_scale_y'><code>SeqPlotR::SeqElement$.infer_scale_y()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqPlotR" data-topic="SeqElement" data-id="collect_legend_keys"><a href='../../SeqPlotR/html/SeqElement.html#method-SeqElement-collect_legend_keys'><code>SeqPlotR::SeqElement$collect_legend_keys()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqPlotR" data-topic="SeqElement" data-id="resolve"><a href='../../SeqPlotR/html/SeqElement.html#method-SeqElement-resolve'><code>SeqPlotR::SeqElement$resolve()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqCurve-new"></a>}}
\if{latex}{\out{\hypertarget{method-SeqCurve-new}{}}}
\subsection{Method \code{new()}}{
Construct a SeqCurveR6.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqCurveR6$new(data = NULL, mapping = NULL, aesthetics = aes(), ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Optional \code{GRanges}.}

\item{\code{mapping}}{Optional \code{SeqMap}. Required: \code{x}, \code{y}, \code{x_end}, \code{y_end}.}

\item{\code{aesthetics}}{Optional \code{SeqAes}. \code{curvature} (default 0.3) sets
the fractional y-offset of the bezier control points.}

\item{\code{...}}{Unused.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqCurve-prep"></a>}}
\if{latex}{\out{\hypertarget{method-SeqCurve-prep}{}}}
\subsection{Method \code{prep()}}{
Resolve mappings, compute bezier grobs per window.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqCurveR6$prep(layout_track, track_windows)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{layout_track}}{Per-window panel metadata list.}

\item{\code{track_windows}}{Track's \code{windows} GRanges.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqCurve-draw"></a>}}
\if{latex}{\out{\hypertarget{method-SeqCurve-draw}{}}}
\subsection{Method \code{draw()}}{
Draw bezier curves via \code{grid::bezierGrob()}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqCurveR6$draw()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqCurve-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SeqCurve-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqCurveR6$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
