% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_density.R
\name{SeqDensityR6}
\alias{SeqDensityR6}
\title{SeqDensity R6 class}
\description{
SeqDensity R6 class

SeqDensity R6 class
}
\details{
Internal R6 generator backing \code{\link[=seq_density]{seq_density()}}. Inherits from
\code{\link{SeqElementR6}}.
}
\keyword{internal}
\section{Super class}{
\code{SeqPlotR::SeqElement} -> \code{SeqDensity}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SeqDensity-new}{\code{SeqDensityR6$new()}}
\item \href{#method-SeqDensity-prep}{\code{SeqDensityR6$prep()}}
\item \href{#method-SeqDensity-draw}{\code{SeqDensityR6$draw()}}
\item \href{#method-SeqDensity-clone}{\code{SeqDensityR6$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SeqPlotR" data-topic="SeqElement" data-id=".infer_scale_y"><a href='../../SeqPlotR/html/SeqElement.html#method-SeqElement-.infer_scale_y'><code>SeqPlotR::SeqElement$.infer_scale_y()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqPlotR" data-topic="SeqElement" data-id="collect_legend_keys"><a href='../../SeqPlotR/html/SeqElement.html#method-SeqElement-collect_legend_keys'><code>SeqPlotR::SeqElement$collect_legend_keys()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqPlotR" data-topic="SeqElement" data-id="resolve"><a href='../../SeqPlotR/html/SeqElement.html#method-SeqElement-resolve'><code>SeqPlotR::SeqElement$resolve()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDensity-new"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDensity-new}{}}}
\subsection{Method \code{new()}}{
Construct a SeqDensityR6.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDensityR6$new(data = NULL, mapping = NULL, aesthetics = aes(), ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Optional \code{GRanges}.}

\item{\code{mapping}}{Optional \code{SeqMap}. Required: \code{y}.}

\item{\code{aesthetics}}{Optional \code{SeqAes}. \code{bw} controls the kernel
bandwidth (passed to \code{\link[stats:density]{stats::density()}}). Default fill \code{"grey60"},
alpha 0.8.}

\item{\code{...}}{Unused.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDensity-prep"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDensity-prep}{}}}
\subsection{Method \code{prep()}}{
Compute \code{stats::density()} of the resolved y, then build a
closed polygon. Density evaluation points are mapped to canvas x via
the panel's \code{yscale}; heights are normalised to \verb{[0, 1]} of panel
height.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDensityR6$prep(layout_track, track_windows)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{layout_track}}{Per-window panel metadata list.}

\item{\code{track_windows}}{Track's \code{windows} GRanges.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDensity-draw"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDensity-draw}{}}}
\subsection{Method \code{draw()}}{
Draw the density polygon via \code{grid::grid.polygon()}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDensityR6$draw()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDensity-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDensity-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDensityR6$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
