% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_line.R
\name{SeqLineR6}
\alias{SeqLineR6}
\title{SeqLine R6 class}
\description{
SeqLine R6 class

SeqLine R6 class
}
\details{
Internal R6 generator backing \code{\link[=seq_line]{seq_line()}}. Inherits from \code{\link{SeqElementR6}}.
}
\keyword{internal}
\section{Super class}{
\code{SeqPlotR::SeqElement} -> \code{SeqLine}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SeqLine-new}{\code{SeqLineR6$new()}}
\item \href{#method-SeqLine-prep}{\code{SeqLineR6$prep()}}
\item \href{#method-SeqLine-draw}{\code{SeqLineR6$draw()}}
\item \href{#method-SeqLine-clone}{\code{SeqLineR6$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SeqPlotR" data-topic="SeqElement" data-id=".infer_scale_y"><a href='../../SeqPlotR/html/SeqElement.html#method-SeqElement-.infer_scale_y'><code>SeqPlotR::SeqElement$.infer_scale_y()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqPlotR" data-topic="SeqElement" data-id="collect_legend_keys"><a href='../../SeqPlotR/html/SeqElement.html#method-SeqElement-collect_legend_keys'><code>SeqPlotR::SeqElement$collect_legend_keys()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqPlotR" data-topic="SeqElement" data-id="resolve"><a href='../../SeqPlotR/html/SeqElement.html#method-SeqElement-resolve'><code>SeqPlotR::SeqElement$resolve()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqLine-new"></a>}}
\if{latex}{\out{\hypertarget{method-SeqLine-new}{}}}
\subsection{Method \code{new()}}{
Construct a SeqLineR6.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqLineR6$new(data = NULL, mapping = NULL, aesthetics = aes(), ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Optional \code{GRanges}.}

\item{\code{mapping}}{Optional \code{SeqMap}.}

\item{\code{aesthetics}}{Optional \code{SeqAes}. \code{type = "step"} (or \code{"s"}) enables
step-line rendering.}

\item{\code{...}}{Unused.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqLine-prep"></a>}}
\if{latex}{\out{\hypertarget{method-SeqLine-prep}{}}}
\subsection{Method \code{prep()}}{
Resolve mappings, clip to windows, transform to npc.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqLineR6$prep(layout_track, track_windows)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{layout_track}}{Per-window panel metadata list.}

\item{\code{track_windows}}{Track's \code{windows} GRanges.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqLine-draw"></a>}}
\if{latex}{\out{\hypertarget{method-SeqLine-draw}{}}}
\subsection{Method \code{draw()}}{
Draw the polyline with \code{grid::grid.lines()}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqLineR6$draw()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqLine-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SeqLine-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqLineR6$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
