% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_lollipop.R
\name{SeqLollipopR6}
\alias{SeqLollipopR6}
\title{SeqLollipop R6 class}
\description{
SeqLollipop R6 class

SeqLollipop R6 class
}
\details{
Internal R6 generator backing \code{\link[=seq_lollipop]{seq_lollipop()}}. Inherits from
\code{\link{SeqElementR6}}.
}
\keyword{internal}
\section{Super class}{
\code{SeqPlotR::SeqElement} -> \code{SeqLollipop}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SeqLollipop-new}{\code{SeqLollipopR6$new()}}
\item \href{#method-SeqLollipop-prep}{\code{SeqLollipopR6$prep()}}
\item \href{#method-SeqLollipop-draw}{\code{SeqLollipopR6$draw()}}
\item \href{#method-SeqLollipop-clone}{\code{SeqLollipopR6$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SeqPlotR" data-topic="SeqElement" data-id=".infer_scale_y"><a href='../../SeqPlotR/html/SeqElement.html#method-SeqElement-.infer_scale_y'><code>SeqPlotR::SeqElement$.infer_scale_y()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqPlotR" data-topic="SeqElement" data-id="collect_legend_keys"><a href='../../SeqPlotR/html/SeqElement.html#method-SeqElement-collect_legend_keys'><code>SeqPlotR::SeqElement$collect_legend_keys()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqPlotR" data-topic="SeqElement" data-id="resolve"><a href='../../SeqPlotR/html/SeqElement.html#method-SeqElement-resolve'><code>SeqPlotR::SeqElement$resolve()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqLollipop-new"></a>}}
\if{latex}{\out{\hypertarget{method-SeqLollipop-new}{}}}
\subsection{Method \code{new()}}{
Construct a SeqLollipopR6.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqLollipopR6$new(data = NULL, mapping = NULL, aesthetics = aes(), ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Optional \code{GRanges}.}

\item{\code{mapping}}{Optional \code{SeqMap}. Recognised: \code{x}, \code{y}.}

\item{\code{aesthetics}}{Optional \code{SeqAes}. \code{baseline} sets the stem's lower
y-value (default 0). Also supports \code{color}, \code{linewidth}, \code{size}.}

\item{\code{...}}{Unused.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqLollipop-prep"></a>}}
\if{latex}{\out{\hypertarget{method-SeqLollipop-prep}{}}}
\subsection{Method \code{prep()}}{
Resolve mappings, clip to windows, transform to npc. Builds
per-window \code{list(x, y0, y1)} used by both the stem and the head.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqLollipopR6$prep(layout_track, track_windows)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{layout_track}}{Per-window panel metadata list.}

\item{\code{track_windows}}{Track's \code{windows} GRanges.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqLollipop-draw"></a>}}
\if{latex}{\out{\hypertarget{method-SeqLollipop-draw}{}}}
\subsection{Method \code{draw()}}{
Draw stems via \code{grid::grid.segments()} and heads via
\code{grid::grid.points()}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqLollipopR6$draw()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqLollipop-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SeqLollipop-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqLollipopR6$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
