% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_poly.R
\name{SeqPolyR6}
\alias{SeqPolyR6}
\title{SeqPoly R6 class}
\description{
SeqPoly R6 class

SeqPoly R6 class
}
\details{
Internal R6 generator backing \code{\link[=seq_poly]{seq_poly()}}. Inherits from \code{\link{SeqElementR6}}.
}
\keyword{internal}
\section{Super class}{
\code{SeqPlotR::SeqElement} -> \code{SeqPoly}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SeqPoly-new}{\code{SeqPolyR6$new()}}
\item \href{#method-SeqPoly-prep}{\code{SeqPolyR6$prep()}}
\item \href{#method-SeqPoly-draw}{\code{SeqPolyR6$draw()}}
\item \href{#method-SeqPoly-clone}{\code{SeqPolyR6$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SeqPlotR" data-topic="SeqElement" data-id=".infer_scale_y"><a href='../../SeqPlotR/html/SeqElement.html#method-SeqElement-.infer_scale_y'><code>SeqPlotR::SeqElement$.infer_scale_y()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqPlotR" data-topic="SeqElement" data-id="collect_legend_keys"><a href='../../SeqPlotR/html/SeqElement.html#method-SeqElement-collect_legend_keys'><code>SeqPlotR::SeqElement$collect_legend_keys()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqPlotR" data-topic="SeqElement" data-id="resolve"><a href='../../SeqPlotR/html/SeqElement.html#method-SeqElement-resolve'><code>SeqPlotR::SeqElement$resolve()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqPoly-new"></a>}}
\if{latex}{\out{\hypertarget{method-SeqPoly-new}{}}}
\subsection{Method \code{new()}}{
Construct a SeqPolyR6.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqPolyR6$new(data = NULL, mapping = NULL, aesthetics = aes(), ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Optional \code{GRanges}.}

\item{\code{mapping}}{Optional \code{SeqMap}. Required: \code{x}, \code{y}. Optional: \code{group}.}

\item{\code{aesthetics}}{Optional \code{SeqAes}.}

\item{\code{...}}{Unused.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqPoly-prep"></a>}}
\if{latex}{\out{\hypertarget{method-SeqPoly-prep}{}}}
\subsection{Method \code{prep()}}{
Resolve mappings, clip, transform, partition by \code{group}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqPolyR6$prep(layout_track, track_windows)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{layout_track}}{Per-window panel metadata list.}

\item{\code{track_windows}}{Track's \code{windows} GRanges.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqPoly-draw"></a>}}
\if{latex}{\out{\hypertarget{method-SeqPoly-draw}{}}}
\subsection{Method \code{draw()}}{
Draw polygons via \code{grid::grid.polygon()} with \code{id}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqPolyR6$draw()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqPoly-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SeqPoly-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqPolyR6$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
