% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_ribbon.R
\name{SeqRibbonR6}
\alias{SeqRibbonR6}
\title{SeqRibbon R6 class}
\description{
SeqRibbon R6 class

SeqRibbon R6 class
}
\details{
Internal R6 generator backing \code{\link[=seq_ribbon]{seq_ribbon()}}. Inherits from
\code{\link{SeqElementR6}}.
}
\keyword{internal}
\section{Super class}{
\code{SeqPlotR::SeqElement} -> \code{SeqRibbon}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SeqRibbon-new}{\code{SeqRibbonR6$new()}}
\item \href{#method-SeqRibbon-prep}{\code{SeqRibbonR6$prep()}}
\item \href{#method-SeqRibbon-draw}{\code{SeqRibbonR6$draw()}}
\item \href{#method-SeqRibbon-clone}{\code{SeqRibbonR6$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SeqPlotR" data-topic="SeqElement" data-id=".infer_scale_y"><a href='../../SeqPlotR/html/SeqElement.html#method-SeqElement-.infer_scale_y'><code>SeqPlotR::SeqElement$.infer_scale_y()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqPlotR" data-topic="SeqElement" data-id="collect_legend_keys"><a href='../../SeqPlotR/html/SeqElement.html#method-SeqElement-collect_legend_keys'><code>SeqPlotR::SeqElement$collect_legend_keys()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqPlotR" data-topic="SeqElement" data-id="resolve"><a href='../../SeqPlotR/html/SeqElement.html#method-SeqElement-resolve'><code>SeqPlotR::SeqElement$resolve()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqRibbon-new"></a>}}
\if{latex}{\out{\hypertarget{method-SeqRibbon-new}{}}}
\subsection{Method \code{new()}}{
Construct a SeqRibbonR6.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqRibbonR6$new(data = NULL, mapping = NULL, aesthetics = aes(), ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Optional \code{GRanges}.}

\item{\code{mapping}}{Optional \code{SeqMap}. Required: \code{x}, \code{y_min}, \code{y_max}.}

\item{\code{aesthetics}}{Optional \code{SeqAes}. Defaults to grey fill, alpha 0.8.}

\item{\code{...}}{Unused.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqRibbon-prep"></a>}}
\if{latex}{\out{\hypertarget{method-SeqRibbon-prep}{}}}
\subsection{Method \code{prep()}}{
Resolve mappings, clip to windows, transform to npc, and
build a closed polygon bounded above by \code{y_max} and below by \code{y_min}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqRibbonR6$prep(layout_track, track_windows)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{layout_track}}{Per-window panel metadata list.}

\item{\code{track_windows}}{Track's \code{windows} GRanges.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqRibbon-draw"></a>}}
\if{latex}{\out{\hypertarget{method-SeqRibbon-draw}{}}}
\subsection{Method \code{draw()}}{
Draw the ribbon polygon via \code{grid::grid.polygon()}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqRibbonR6$draw()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqRibbon-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SeqRibbon-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqRibbonR6$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
