% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_segment.R
\name{SeqSegmentR6}
\alias{SeqSegmentR6}
\title{SeqSegment R6 class}
\description{
SeqSegment R6 class

SeqSegment R6 class
}
\details{
Internal R6 generator backing \code{\link[=seq_segment]{seq_segment()}}. Inherits from
\code{\link{SeqElementR6}}.
}
\keyword{internal}
\section{Super class}{
\code{SeqPlotR::SeqElement} -> \code{SeqSegment}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SeqSegment-new}{\code{SeqSegmentR6$new()}}
\item \href{#method-SeqSegment-prep}{\code{SeqSegmentR6$prep()}}
\item \href{#method-SeqSegment-draw}{\code{SeqSegmentR6$draw()}}
\item \href{#method-SeqSegment-clone}{\code{SeqSegmentR6$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SeqPlotR" data-topic="SeqElement" data-id=".infer_scale_y"><a href='../../SeqPlotR/html/SeqElement.html#method-SeqElement-.infer_scale_y'><code>SeqPlotR::SeqElement$.infer_scale_y()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqPlotR" data-topic="SeqElement" data-id="collect_legend_keys"><a href='../../SeqPlotR/html/SeqElement.html#method-SeqElement-collect_legend_keys'><code>SeqPlotR::SeqElement$collect_legend_keys()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqPlotR" data-topic="SeqElement" data-id="resolve"><a href='../../SeqPlotR/html/SeqElement.html#method-SeqElement-resolve'><code>SeqPlotR::SeqElement$resolve()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqSegment-new"></a>}}
\if{latex}{\out{\hypertarget{method-SeqSegment-new}{}}}
\subsection{Method \code{new()}}{
Construct a SeqSegmentR6.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqSegmentR6$new(data = NULL, mapping = NULL, aesthetics = aes(), ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Optional \code{GRanges}.}

\item{\code{mapping}}{Optional \code{SeqMap}. Required fields: \code{x}, \code{x_end}, \code{y},
\code{y_end}.}

\item{\code{aesthetics}}{Optional \code{SeqAes}.}

\item{\code{...}}{Unused.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqSegment-prep"></a>}}
\if{latex}{\out{\hypertarget{method-SeqSegment-prep}{}}}
\subsection{Method \code{prep()}}{
Resolve mappings, apply auto-scale for non-concrete color,
clip to windows, transform to npc.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqSegmentR6$prep(layout_track, track_windows)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{layout_track}}{Per-window panel metadata list.}

\item{\code{track_windows}}{Track's \code{windows} GRanges.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqSegment-draw"></a>}}
\if{latex}{\out{\hypertarget{method-SeqSegment-draw}{}}}
\subsection{Method \code{draw()}}{
Draw segments via \code{grid::grid.segments()}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqSegmentR6$draw()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqSegment-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SeqSegment-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqSegmentR6$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
