% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_text.R
\name{SeqTextR6}
\alias{SeqTextR6}
\title{SeqText R6 class}
\description{
SeqText R6 class

SeqText R6 class
}
\details{
Internal R6 generator backing \code{\link[=seq_text]{seq_text()}}. Inherits from \code{\link{SeqElementR6}}.
}
\keyword{internal}
\section{Super class}{
\code{SeqPlotR::SeqElement} -> \code{SeqText}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SeqText-new}{\code{SeqTextR6$new()}}
\item \href{#method-SeqText-prep}{\code{SeqTextR6$prep()}}
\item \href{#method-SeqText-draw}{\code{SeqTextR6$draw()}}
\item \href{#method-SeqText-clone}{\code{SeqTextR6$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SeqPlotR" data-topic="SeqElement" data-id=".infer_scale_y"><a href='../../SeqPlotR/html/SeqElement.html#method-SeqElement-.infer_scale_y'><code>SeqPlotR::SeqElement$.infer_scale_y()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqPlotR" data-topic="SeqElement" data-id="collect_legend_keys"><a href='../../SeqPlotR/html/SeqElement.html#method-SeqElement-collect_legend_keys'><code>SeqPlotR::SeqElement$collect_legend_keys()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqPlotR" data-topic="SeqElement" data-id="resolve"><a href='../../SeqPlotR/html/SeqElement.html#method-SeqElement-resolve'><code>SeqPlotR::SeqElement$resolve()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqText-new"></a>}}
\if{latex}{\out{\hypertarget{method-SeqText-new}{}}}
\subsection{Method \code{new()}}{
Construct a SeqTextR6.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqTextR6$new(data = NULL, mapping = NULL, aesthetics = aes(), ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Optional \code{GRanges}.}

\item{\code{mapping}}{Optional \code{SeqMap}. Required: \code{x}, \code{y}, \code{label}.}

\item{\code{aesthetics}}{Optional \code{SeqAes}. Supports \code{size}, \code{color}, \code{angle},
\code{hjust}, \code{vjust}, and a constant \code{label}.}

\item{\code{...}}{Unused.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqText-prep"></a>}}
\if{latex}{\out{\hypertarget{method-SeqText-prep}{}}}
\subsection{Method \code{prep()}}{
Resolve mappings, clip, transform to npc.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqTextR6$prep(layout_track, track_windows)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{layout_track}}{Per-window panel metadata list.}

\item{\code{track_windows}}{Track's \code{windows} GRanges.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqText-draw"></a>}}
\if{latex}{\out{\hypertarget{method-SeqText-draw}{}}}
\subsection{Method \code{draw()}}{
Draw text via \code{grid::grid.text()}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqTextR6$draw()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqText-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SeqText-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqTextR6$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
